Close file before deleting.
Travis prepare

Travis packages without sudo

travis

travis packge edit

travis packge edit

Travis csript

Travis debug

Test compiler

Added packages libc6

Added packages 32 bit gcc multilib

Linux headers package

Linux headers

Linux headers

Build ess

Linux headers

Linux headers

Linux headers

GCC

GCC

GCC

GCC

Arch export

Arch export

Headers version

Arch export

Headers export

Headers export

C

+++ bulldog-board-raspberrypi/src/main/java/io/silverspoon/bulldog/raspberrypi/pwm/PwmFrequencyCalculator.java
@@ -31,7 +31,7 @@ public class PwmFrequencyCalculator {
          double divisor = CLOCK_FREQUENCY / frequency;
          int DIVI = (int) Math.floor(divisor);
          int DIVF = (int) Math.floor((divisor - DIVI) * 1024);
-         int divreg = (0x5a << 24) | ((int) DIVI << 12) | (DIVF);
+         int divreg = (0x5a << 24) | (DIVI << 12) | (DIVF);
 
          if (Double.isNaN(DIVF) || Double.isInfinite(DIVI) || DIVI < 1 || DIVI > 4095) {
             error = true;
